+++ spring-batch-core/src/main/resources/org/springframework/batch/core/schema-hana.sql
@@ -19,7 +19,6 @@ CREATE TABLE BATCH_JOB_EXECUTION  (
 	EXIT_CODE VARCHAR(2500) ,
 	EXIT_MESSAGE VARCHAR(2500) ,
 	LAST_UPDATED TIMESTAMP,
-	JOB_CONFIGURATION_LOCATION VARCHAR(2500) ,
 	constraint JOB_INST_EXEC_FK foreign key (JOB_INSTANCE_ID)
 	references BATCH_JOB_INSTANCE(JOB_INSTANCE_ID)
 ) ;
