--- spring-batch-core/src/main/sql/hana.properties
@@ -1,14 +1,0 @@
-platform=hana
-# SQL language oddities
-BIGINT = BIGINT
-IDENTITY = 
-GENERATED = GENERATED BY DEFAULT AS IDENTITY
-IFEXISTSBEFORE = 
-DOUBLE = DOUBLE
-BLOB = BLOB
-CLOB = CLOB
-TIMESTAMP = TIMESTAMP
-VARCHAR = VARCHAR
-CHAR = VARCHAR
-# for generating drop statements...
-SEQUENCE = SEQUENCE
