+++ spring-batch-infrastructure/src/main/java/org/springframework/batch/item/database/support/HanaPagingQueryProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2018 the original author or authors.
+ * Copyright 2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ import org.springframework.util.StringUtils;
  * SAP HANA implementation of a {@link PagingQueryProvider} using database specific features.
  *
  * @author Jonathan Bregler
- * @since 4.0
+ * @since 5.0
  */
 public class HanaPagingQueryProvider extends AbstractSqlPagingQueryProvider {
 
@@ -35,7 +35,7 @@ public class HanaPagingQueryProvider extends AbstractSqlPagingQueryProvider {
 	@Override
 	public String generateRemainingPagesQuery(int pageSize) {
 		if(StringUtils.hasText(getGroupClause())) {
-			return SqlPagingQueryUtils.generateLimitGroupedSqlQuery(this, true, buildLimitClause(pageSize));
+			return SqlPagingQueryUtils.generateLimitGroupedSqlQuery(this, buildLimitClause(pageSize));
 		}
 		else {
 			return SqlPagingQueryUtils.generateLimitSqlQuery(this, true, buildLimitClause(pageSize));
