+++ spring-batch-infrastructure/src/test/java/org/springframework/batch/item/database/support/HanaPagingQueryProviderTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2018 the original author or authors.
+ * Copyright 2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,7 +26,7 @@ import static org.junit.Assert.assertEquals;
 
 /**
  * @author Jonathan Bregler
- * @since 4.0
+ * @since 5.0
  */
 public class HanaPagingQueryProviderTests extends AbstractSqlPagingQueryProviderTests {
