--- spring-batch-samples/src/main/sql/hana.vpp
@@ -1,2 +1,0 @@
-#macro (sequence $name $value)CREATE SEQUENCE ${name} START WITH ${value} MINVALUE 0;
-#end
